linux-gen: pktio: move pktio socket stats to a new file
Cleanup pktio code by moving socket stats functions to a
new file.
