+++ platform/linux-generic/pktio/netmap.c
@@ -16,6 +16,7 @@
 #include <odp/api/plat/time_inlines.h>
 
 #include <odp_packet_io_internal.h>
+#include <odp_packet_io_stats.h>
 #include <odp_ethtool_stats.h>
 #include <odp_ethtool_rss.h>
 #include <odp_socket_common.h>
