+++ platform/linux-generic/pktio/socket.c
@@ -39,8 +39,7 @@
 #include <odp_socket_common.h>
 #include <odp_packet_internal.h>
 #include <odp_packet_io_internal.h>
-#include <odp_ethtool_stats.h>
-#include <odp_sysfs_stats.h>
+#include <odp_packet_io_stats.h>
 #include <odp_align_internal.h>
 #include <odp_debug_internal.h>
 #include <odp_errno_define.h>
@@ -148,7 +147,6 @@ static int sock_setup_pkt(pktio_entry_t *pktio_entry, const char *netdev,
 	struct sockaddr_ll sa_ll;
 	char shm_name[ODP_SHM_NAME_LEN];
 	pkt_sock_t *pkt_sock = pkt_priv(pktio_entry);
-	odp_pktio_stats_t cur_stats;
 
 	/* Init pktio entry */
 	memset(pkt_sock, 0, sizeof(*pkt_sock));
@@ -200,21 +198,10 @@ static int sock_setup_pkt(pktio_entry_t *pktio_entry, const char *netdev,
 		goto error;
 	}
 
-	err = ethtool_stats_get_fd(pkt_priv(pktio_entry)->sockfd,
-				   pktio_entry->s.name,
-				   &cur_stats);
-	if (err != 0) {
-		err = sysfs_stats(pktio_entry, &cur_stats);
-		if (err != 0) {
-			pktio_entry->s.stats_type = STATS_UNSUPPORTED;
-			ODP_DBG("pktio: %s unsupported stats\n",
-				pktio_entry->s.name);
-		} else {
-		pktio_entry->s.stats_type = STATS_SYSFS;
-		}
-	} else {
-		pktio_entry->s.stats_type = STATS_ETHTOOL;
-	}
+	pktio_entry->s.stats_type = sock_stats_type_fd(pktio_entry,
+						       pkt_sock->sockfd);
+	if (pktio_entry->s.stats_type == STATS_UNSUPPORTED)
+		ODP_DBG("pktio: %s unsupported stats\n", pktio_entry->s.name);
 
 	err = sock_stats_reset(pktio_entry);
 	if (err != 0)
